Extract Cursor to separate files.
