+++ Makefile
@@ -1,5 +1,5 @@
 Target = cod
-Sources = ./src/cod.cpp
+Sources = ./src/cursor.cpp ./src/cod.cpp
 CXX = g++
 CXXFLAGS = -O3
 DESTPATH = ./bin/$(Target)
