+++ src/cod.cpp
@@ -2,6 +2,8 @@
 #include <termios.h>
 #include <unistd.h>
 
+#include "cursor.h"
+
 using namespace std;
 
 const string ESC_CLEAR_SCREEN = "\e[2J";
@@ -15,56 +17,6 @@ class Buffer {
 
     void clear() {
       content = "";
-    }
-};
-
-class Cursor {
-  public:
-    Cursor(int n, int m) {
-      cy = n;
-      cx = m;
-    }
-    Cursor(int n) {
-      cy = n;
-      cx = 1;
-    }
-    Cursor() {
-      cy = 1;
-      cx = 1;
-    }
-
-    int cy;
-    int cx;
-
-    void down() {
-      cy++;
-    }
-    void inspect() {
-      cout << " cx: " << cx << " cy: " << cy << endl;
-    }
-    void left() {
-      if (cx <= 0) {
-        cx = 1;
-      } else if (cx - 1 < 1) {
-        cx = 1;
-      } else {
-        cx--;
-      }
-    }
-    void render() {
-      cout << "\e[" << cy << ";" << cx << "H";
-    }
-    void right() {
-      cx++;
-    }
-    void up() {
-      if (cy <= 0) {
-        cy = 1;
-      } else if (cy - 1 < 1) {
-        cy = 1;
-      } else {
-        cy--;
-      }
     }
 };
